backend/requirements.txt
fastapi
uvicorn
langchain
google-api-python-client
python-dotenv
google-auth-oauthlib
langchain-google-genai